split out rescue team and text util stuff
