+++ src/save.c
@@ -74,7 +74,7 @@ extern u32 gUnknown_203B17C;
 extern u32 gUnknown_203B180;
 extern u32 *gUnknown_203B45C;
 extern u32 *gUnknown_203B460;
-extern u32 gRescueTeamInfoRef;
+extern struct RescueTeamData *gRescueTeamInfoRef;
 extern u8 *gFriendAreas;
 extern u32 gUnknown_203B46C;
 extern u8 *gUnknown_203B480;
@@ -116,7 +116,6 @@ extern u32 sub_808F2B0(void* a, s32 b);
 extern u32 sub_8091C68(u8 *, u32);
 extern u32 sub_8091D14(void* a, s32 b);
 extern u32 sub_80921C4(u8 *, u32);
-extern u32 sub_809222C(void* a, s32 b);
 extern u32 sub_80927A8(u8 *, u32);
 extern u32 sub_80927F4(void* a, s32 b);
 extern u8 *sub_8095100(void);
@@ -136,8 +135,6 @@ extern u32 *sub_808CE00(void);
 extern void sub_808CE08(void);
 extern u32 *sub_80909D0(void);
 extern void sub_80909D8(void);
-extern u32 GetRescueTeamInfo(void);
-extern void InitializeResuceTeamInfo(void);
 extern u32 sub_8094990(void);
 extern void sub_8094998(u8 r0);
 extern u8 *sub_80950F8(void);
